add tree api add loading
