+++ web/src/dtos/role-user-permissions/index.ts
@@ -72,3 +72,23 @@ export interface IRolePermissionDataItem {
 export interface IDeleteRole {
   roleIds: string[];
 }
+
+export interface DepartmentType {
+  id: string;
+  name: string;
+  parentId: string;
+}
+
+export interface UsersDto {
+  id: string;
+  userName: string;
+}
+
+export interface DepartmentTreeDto {
+  department: DepartmentType;
+  staffs: UsersDto[] | [];
+  childrens: DepartmentTreeDto[];
+}
+export interface IFoundationTreeDto {
+  staffDepartmentHierarchy: DepartmentTreeDto[];
+}
