+++ web/src/pages/role-perimission/components/role-from/index.tsx
@@ -6,12 +6,13 @@ import styles from "./index.module.scss";
 import ArrowRightIcon from "@mui/icons-material/ArrowRight";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import IndeterminateCheckBoxIcon from "@mui/icons-material/IndeterminateCheckBox";
-
+import { v4 as uuidv4 } from "uuid";
 import {
   Autocomplete,
   Button,
   Card,
   Checkbox,
+  CircularProgress,
   FormControlLabel,
   Input,
   Stack,
@@ -34,6 +35,7 @@ export const RoleFrom = () => {
     removeOption,
     rolePermissionsCheckedList,
     role,
+    isLoading,
     updateRole,
     addOrModifyRolePermission,
     handleUpdateRolePermissionsChecked,
@@ -42,6 +44,8 @@ export const RoleFrom = () => {
   const icon = <CheckBoxOutlineBlankIcon fontSize="small" />;
 
   const checkedIcon = <CheckBoxIcon fontSize="small" />;
+
+  const isShowLoading = isLoading && location.pathname.includes("/roles/edit");
 
   const renderInputBox = (
     title: string,
@@ -90,6 +94,7 @@ export const RoleFrom = () => {
                     display: "flex",
                     alignItems: "center",
                   }}
+                  key={uuidv4()}
                 >
                   <div
                     style={{
@@ -194,85 +199,103 @@ export const RoleFrom = () => {
         </div>
 
         <Card className={styles.card} variant="outlined">
-          <Stack spacing={3}>
-            <div className={styles.itemTitle}>角色信息</div>
-            <div className={styles.item}>
-              <div className={styles.itemSubTitle}>角色名稱：</div>
-              <div className={styles.itemInput}>
-                <Input
-                  disableUnderline={true}
-                  sx={inputStyles}
-                  placeholder="請輸入角色名稱"
-                  value={role?.name ?? ""}
-                  onChange={(e) => updateRole("name", e.target.value)}
-                />
-              </div>
+          {isShowLoading ? (
+            <div style={{ textAlign: "center" }}>
+              <CircularProgress />
             </div>
-            <div className={styles.item}>
-              <div className={styles.itemSubTitle}>角色描述：</div>
-              <div className={styles.itemInput}>
-                <Input
-                  disableUnderline={true}
-                  sx={inputStyles}
-                  placeholder="請輸入角色描述"
-                  value={role?.description ?? ""}
-                  onChange={(e) => updateRole("description", e.target.value)}
-                />
-              </div>
-            </div>
-          </Stack>
-        </Card>
-
-        <Card className={styles.card} variant="outlined">
-          <Stack spacing={3}>
-            <div className={styles.itemTitle}>功能權限</div>
-            <div className={styles.item}>
-              <div className={styles.itemPermission}>
-                <div className={styles.itemPerssionsForm}>
-                  {rolePermissionsCheckedList.map((roleItem, index) => {
-                    return (
-                      <div style={{ marginRight: 15 }} key={index}>
-                        <FormControlLabel
-                          key={index}
-                          control={
-                            <Checkbox
-                              onChange={(e) =>
-                                handleUpdateRolePermissionsChecked(
-                                  roleItem.id,
-                                  e.target.checked
-                                )
-                              }
-                              checked={roleItem.checked}
-                            />
-                          }
-                          label={roleItem.name}
-                        />
-                      </div>
-                    );
-                  })}
+          ) : (
+            <Stack spacing={3}>
+              <div className={styles.itemTitle}>角色信息</div>
+              <div className={styles.item}>
+                <div className={styles.itemSubTitle}>角色名稱：</div>
+                <div className={styles.itemInput}>
+                  <Input
+                    disableUnderline={true}
+                    sx={inputStyles}
+                    placeholder="請輸入角色名稱"
+                    value={role?.name ?? ""}
+                    onChange={(e) => updateRole("name", e.target.value)}
+                  />
                 </div>
               </div>
-            </div>
-          </Stack>
+              <div className={styles.item}>
+                <div className={styles.itemSubTitle}>角色描述：</div>
+                <div className={styles.itemInput}>
+                  <Input
+                    disableUnderline={true}
+                    sx={inputStyles}
+                    placeholder="請輸入角色描述"
+                    value={role?.description ?? ""}
+                    onChange={(e) => updateRole("description", e.target.value)}
+                  />
+                </div>
+              </div>
+            </Stack>
+          )}
         </Card>
 
         <Card className={styles.card} variant="outlined">
-          <Stack spacing={3}>
-            <div className={styles.itemTitle}>數據權限</div>
-            {renderInputBox(
-              "拉群功能：",
-              "pullCrowdData",
-              checkboxData.pullCrowdData,
-              showLabel.pullCrowdData
-            )}
-            {renderInputBox(
-              "通知功能：",
-              "notificationData",
-              checkboxData.notificationData,
-              showLabel.notificationData,
-              true
-            )}
-          </Stack>
+          {isShowLoading ? (
+            <div style={{ textAlign: "center" }}>
+              <CircularProgress />
+            </div>
+          ) : (
+            <Stack spacing={3}>
+              <div className={styles.itemTitle}>功能權限</div>
+              <div className={styles.item}>
+                <div className={styles.itemPermission}>
+                  <div className={styles.itemPerssionsForm}>
+                    {rolePermissionsCheckedList.map((roleItem, index) => {
+                      return (
+                        <div style={{ marginRight: 15 }} key={index}>
+                          <FormControlLabel
+                            key={index}
+                            control={
+                              <Checkbox
+                                onChange={(e) =>
+                                  handleUpdateRolePermissionsChecked(
+                                    roleItem.id,
+                                    e.target.checked
+                                  )
+                                }
+                                checked={roleItem.checked}
+                              />
+                            }
+                            label={roleItem.name}
+                          />
+                        </div>
+                      );
+                    })}
+                  </div>
+                </div>
+              </div>
+            </Stack>
+          )}
+        </Card>
+
+        <Card className={styles.card} variant="outlined">
+          {isShowLoading ? (
+            <div style={{ textAlign: "center" }}>
+              <CircularProgress />
+            </div>
+          ) : (
+            <Stack spacing={3}>
+              <div className={styles.itemTitle}>數據權限</div>
+              {renderInputBox(
+                "拉群功能：",
+                "pullCrowdData",
+                checkboxData.pullCrowdData,
+                showLabel.pullCrowdData
+              )}
+              {renderInputBox(
+                "通知功能：",
+                "notificationData",
+                checkboxData.notificationData,
+                showLabel.notificationData,
+                true
+              )}
+            </Stack>
+          )}
         </Card>
       </Stack>
     </div>
