+++ web/src/pages/role-perimission/components/role-from/props.ts
@@ -14,23 +14,6 @@ export type AllDepartmentData = {
   notificationData: DepartmentDto[];
 };
 
-export interface DepartmentType {
-  id: string;
-  name: string;
-  parentId: string;
-}
-
-export interface UsersDto {
-  id: string;
-  userName: string;
-}
-
-export interface DepartmentTreeDto {
-  department: DepartmentType;
-  staffs: UsersDto[] | [];
-  childrens: DepartmentTreeDto[];
-}
-
 export interface RolePermissionsDto {
   id: string;
   createdDate: string;
